main_requirements.txt
absl-py==1.3.0
astor==0.8.1
backcall==0.2.0
beautifulsoup4==4.11.1
blis==0.7.9
brotlipy==0.7.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
clip==1.0
colorama==0.4.4
coloredlogs==15.0.1
commonmark==0.9.1
conda-content-trust==0+unknown
conda-package-handling==1.8.1
confection==0.0.3
cryptography==36.0.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.5.1
decorator==5.1.1
en-core-web-sm==3.4.1
en-core-web-trf==3.4.1
entrypoints==0.4
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.38.0
ftfy==6.1.1
gast==0.5.3
gdown==4.5.3
gensim==3.7.3
grammaregex==0.1.3
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.11.0
humanfriendly==10.0
idna==3.3
imageio==2.22.4
importlib-metadata==5.0.0
ipykernel==6.15.2
ipython==7.31.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.2.2
jupyter_core==4.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==2.2.5
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlxtend==0.16.0
mock==4.0.3
mpmath==1.2.1
murmurhash==1.0.9
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.7
numpy==1.21.6
onnx==1.12.0
onnx-simplifier==0.4.8
onnxruntime==1.13.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathy==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.2
preshed==3.0.8
prompt-toolkit==3.0.20
protobuf==4.21.9
psutil==5.8.0
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.10.31
requests==2.27.1
responsibly==0.1.3
rich==12.6.0
ruamel-yaml-conda==0.15.100
scikit-image==0.18.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
srsly==2.4.5
sympy==1.10.1
tabulate==0.9.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
termcolor==2.1.1
textpipeliner==0.3.1
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.21.3
typer==0.7.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wordcloud==1.8.2.2
zipp==3.10.0
